Moved: moved email code to emailSender.js, consolidated code into sendNotification function
