+++ server/server.js
@@ -1,13 +1,12 @@
-const express = require('express');
-const { ApolloServer } = require('apollo-server-express');
-const path = require('path');
-const { authMiddleware } = require('./utils/auth');
-const nodemailer = require('nodemailer');
-const { google } = require("googleapis");
-const OAuth2 = google.auth.OAuth2;
-const { typeDefs, resolvers } = require('./schema');
-const db = require('./config/connection');
-require('dotenv').config();
+const express = require("express");
+const { ApolloServer } = require("apollo-server-express");
+const path = require("path");
+const { authMiddleware } = require("./utils/auth");
+//nodemailer code function
+const { sendEmailNotification } = require('./email/emailSender');
+const { typeDefs, resolvers } = require("./schema");
+const db = require("./config/connection");
+require("dotenv").config();
 
 //Port for the server
 const PORT = process.env.PORT || 3001;
@@ -22,82 +21,38 @@ const server = new ApolloServer({
 app.use(express.urlencoded({ extended: false }));
 app.use(express.json());
 
-if (process.env.NODE_ENV === 'production') {
-  app.use(express.static(path.join(__dirname, '../client/build')));
+if (process.env.NODE_ENV === "production") {
+  app.use(express.static(path.join(__dirname, "../client/build")));
 }
 
-app.get('/', (req, res) => {
-  res.sendFile(path.join(__dirname, '../client/build/index.html'));
+app.get("/", (req, res) => {
+  res.sendFile(path.join(__dirname, "../client/build/index.html"));
 });
 //Apollo server middleware
 const startApolloServer = async () => {
   await server.start();
   server.applyMiddleware({ app });
-  
-  db.once('open', () => {
+
+  db.once("open", () => {
     app.listen(PORT, () => {
       console.log(`API server running on port ${PORT}!`);
-      console.log(`Use GraphQL at http://localhost:${PORT}${server.graphqlPath}`);
-    })
-  })
-
-  // Nodemailer Setup
-  const createTransporter = async () => {
-    const oauth2Client = new OAuth2(
-      process.env.CLIENT_ID,
-      process.env.CLIENT_SECRET,
-      "https://developers.google.com/oauthplayground"
-    );
-  
-    oauth2Client.setCredentials({
-      refresh_token: process.env.REFRESH_TOKEN
+      console.log(
+        `Use GraphQL at http://localhost:${PORT}${server.graphqlPath}`
+      );
     });
-  
-    try {
-      const accessToken = await new Promise((resolve, reject) => {
-        oauth2Client.getAccessToken((err, token) => {
-          if (err) {
-            reject(err); // Pass the error as a reason for rejection
-          }
-          resolve(token);
-        });
-      });
-  
-      const transporter = nodemailer.createTransport({
-        service: "gmail",
-        auth: {
-          type: "OAuth2",
-          user: process.env.EMAIL,
-          accessToken,
-          clientId: process.env.CLIENT_ID,
-          clientSecret: process.env.CLIENT_SECRET,
-          refreshToken: process.env.REFRESH_TOKEN
-        }
-      });
-  
-      return transporter;
-    } catch (error) {
-      throw new Error("Failed to create transporter: " + error.message);
-    }
-  };
-  
-  const sendEmail = async (emailOptions) => {
-    try {
-      let emailTransporter = await createTransporter();
-      await emailTransporter.sendMail(emailOptions);
-    } catch (error) {
-      console.error("Error sending email:", error.message);
-    }
-  };
-  
-  sendEmail({
-    subject: "Test",
-    text: "I am sending an email from nodemailer!",
-    to: "keasiley@icloud.com",
-    from: process.env.EMAIL
   });
-  
+};
+// Send an email notification
+const emailOptions = {
+  subject: "Test",
+  text: "I am sending an email from nodemailer!",
+  to: "keasiley@icloud.com",
+  from: process.env.EMAIL,
+};
 
-  };
+sendEmailNotification(emailOptions);
+
+
+
 //Start Apollo server
 startApolloServer();
